requirements.txt
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.5
celery==5.2.7
redis==4.5.1
hiredis==2.2.2
boto3==1.26.71
django-environ==0.9.0
Pillow==9.4.0
django-cors-headers==3.13.0
shortuuid==1.0.11